LayoutTests/fast/multicol/widows.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<script>
    description("Test column balancer behavior when widows requirements are high, and there's need for an early break to honor widows");
</script>
<p>The 5 lines should all be in the second column.</p>
<div id="mc" style="-webkit-columns:3; columns:3; -webkit-column-rule:solid; line-height:50px; orphans:1; widows:5;">
    <span id="inFirstColumn"><br></span>
    <span id="first">line<br></span>
    line<br>
    line<br>
    line<br>
    <span id="last">line<br></span>
</div>
<script>
    shouldBe("mc.offsetHeight", "250");
    shouldBe("first.offsetTop", "inFirstColumn.offsetTop");
    shouldBeGreaterThan("first.offsetLeft", "inFirstColumn.offsetLeft");
    shouldBe("last.offsetLeft", "first.offsetLeft");
</script>
