third_party/blink/web_tests/fast/multicol/widows-expected.txt

Test column balancer behavior when widows requirements are high, and there's need for an early break to honor widows

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS mc.offsetHeight is 250
PASS first.offsetTop is inFirstColumn.offsetTop
PASS first.offsetLeft is > inFirstColumn.offsetLeft
PASS last.offsetLeft is first.offsetLeft
PASS successfullyParsed is true

TEST COMPLETE


The 5 lines should all be in the second column.


line
line
line
line
line
